+++ example/index.js
@@ -4,8 +4,8 @@ import './example';
 // ------------------------------Обычный селект--------------------
 const dropdown = new CGSelect({
   selector: '.cg-dropdown_one',
-  placeholder: 'Выберите авто',
-  lable: 'EXAMPLE',
+  placeholder: 'Choose a car',
+  label: 'EXAMPLE',
   items: [
     'BMW',
     {
@@ -27,10 +27,14 @@ const dropdown = new CGSelect({
   },
 });
 
+dropdown.on('clear', function (e) {
+  console.log(`this state: ${e}`);
+});
+
 // ------------------------------NativeSelect-----------------------
 const dropdownNativeSelect = new CGSelect({
   selector: '.cg-dropdown_selectNative',
-  placeholder: 'Выберите авто',
+  placeholder: 'Choose a car',
   nativeSelectMode: true,
   items: [
     'BMW',
@@ -53,12 +57,13 @@ const dropdownNativeSelect = new CGSelect({
       width: '824px',
     },
   },
+  theme: 'dark',
 });
 
 // ------------------------------listDisplayMode--------------------
 const dropdownlistDisplayMode = new CGSelect({
   selector: '.cg-dropdown_listDisplayMode',
-  placeholder: 'Выберите авто',
+  placeholder: 'Choose a car',
   listDisplayMode: true,
   items: [
     'BMW',
@@ -84,7 +89,7 @@ const dropdownlistDisplayMode = new CGSelect({
 // --------------------------------Категории--------------------------
 const dropdown4 = new CGSelect({
   selector: '.cg-dropdown_categories',
-  placeholder: 'Выберите регион',
+  placeholder: 'Choose region',
   searchMode: true,
   items: [
     {
@@ -129,7 +134,7 @@ const dropdown4 = new CGSelect({
 //----------------управление с помощью кнопок----------------------------------
 const dropdownBtn = new CGSelect({
   selector: '.cg-dropdown_usedBtn',
-  placeholder: 'Выберите авто',
+  placeholder: 'Choose a car',
   searchMode: true,
   items: [
     'BMW',
@@ -176,7 +181,7 @@ dropdownBtn.buttonControl(buttonClose, 'close');
 //-------------------------Функция Disabled----------------------------------
 const dropdownDisabled = new CGSelect({
   selector: '.cg-dropdown_checkboxDisable',
-  placeholder: 'Выберите авто',
+  placeholder: 'Choose a car',
   searchMode: true,
   items: [
     'BMW',